ci: add GitHub Actions CI and testing infrastructure
Establishes CI pipeline (fmt, clippy -D warnings, test, typos) and
adds E2E testing layers: example smoke tests running real dataflows
through the WS control plane, and real-dataflow WS E2E tests verifying
the full coordinator+daemon+node lifecycle.

Fixes all pre-existing clippy warnings to enable -D warnings in CI.

+++ apis/python/operator/src/lib.rs
@@ -292,7 +292,7 @@ pub fn metadata_to_pydict<'a>(
 
     // Extract seconds and microseconds (Python datetime supports microsecond precision)
     let seconds = duration_since_epoch.as_secs() as i64;
-    let microseconds = duration_since_epoch.subsec_micros() as u32;
+    let microseconds = duration_since_epoch.subsec_micros();
 
     // Get UTC timezone from Python's datetime module and create timezone-aware datetime
     // We use Python's datetime.fromtimestamp() to create a UTC-aware datetime object
